Added Meredith's latest work and moved helper functions and useState from App to Home.
