+++ client/src/App.js
@@ -1,9 +1,8 @@
-import React, {useState, useCallback, useEffect} from 'react';
+import React, {useState, useEffect} from 'react';
 import { BrowserView, MobileView, isBrowser, isMobile } from 'react-device-detect';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import axios from 'axios'
 import ReactDOM from 'react-dom';
-import moment from 'moment';
 import Registration from './components/authentication/Registrationv2.jsx';
 import Login from './components/authentication/Login.jsx';
 import Home from './components/Home.jsx';
@@ -11,38 +10,6 @@ import { example } from './../../database/example.js'
 
 function App() {
   const [currentPage, changePage] = useState("home");
-  const [myEvents, setMyEvents] = useState([
-    {
-      id: 0,
-      title: "Sample Event",
-      start_date: new Date(),
-      end_date: new Date(moment().add(1, "hour")),
-      allDay: false,
-      in_calendar: true
-    },
-    {
-      id: 1,
-      title: 'Trip to China',
-      description: '5-day business trip to meet with manufacturers',
-      duration: '5 days',
-      start: new Date(),
-      end: new Date(moment().add(5, "days")),
-      category_id: 1,
-      in_calendar: false
-    },
-    {
-      id: 2,
-      title: 'Trip to Los Angeles',
-      description: 'Meeting with Executives',
-      duration: '6 hours',
-      start: new Date(moment('14 Apr 2022 09:30:00 UT')),
-      end: new Date(moment('14 Apr 2022 15:30:00 UT')),
-      category_id: 1,
-      in_calendar: false
-    }
-  ]);
-  const [onCalendar, setOnCalendar] = useState(false);
-  const [draggedEvent, setDraggedEvent] = useState()
   const [isLoggedIn, setIsLoggedIn] = useState(false);
   const [isLoading, setIsLoading] = useState(true);
   const [userEmail, setEmail] = useState(null);
@@ -72,72 +39,6 @@ function App() {
     })
   }, [userEmail])
 
-  //Calendar helper functions
-  const moveEvent = useCallback(
-    ({ event, start, end, isAllDay: droppedOnAllDaySlot = false }) => {
-      const { allDay } = event;
-      if (!allDay && droppedOnAllDaySlot) {
-        event.allDay = true;
-      }
-
-      setMyEvents((prev) => {
-        const existing = prev.find((ev) => ev.id === event.id) ?? {};
-        const filtered = prev.filter((ev) => ev.id !== event.id);
-        return [...filtered, { ...existing, start, end, allDay }];
-      });
-    },
-    [setMyEvents]
-  );
-
-  const newEvent = useCallback(
-    (event) => {
-      setMyEvents((prev) => {
-        const idList = prev.map((item) => item.id)
-        const newId = Math.max(...idList) + 1
-        return [...prev, { ...event, id: newId }]
-      })
-    },
-    [setMyEvents]
-  )
-
-  const resizeEvent = useCallback(
-    ({ event, start, end }) => {
-      setMyEvents((prev) => {
-        const existing = prev.find((ev) => ev.id === event.id) ?? {};
-        const filtered = prev.filter((ev) => ev.id !== event.id);
-        return [...filtered, { ...existing, start, end }];
-      });
-    },
-    [setMyEvents]
-  );
-
-  const changeTitle = (event) => {
-    var title = prompt("Change title", event.title);
-    var newList = myEvents;
-    setMyEvents((prev) => {
-      newList[prev.indexOf(event)].title = title;
-      return newList
-    });
-  };
-  const handleDragStart = useCallback((event) => {
-    console.log('dragged event', event)
-    setDraggedEvent(event), []
-  })
-
-
-  const onDropFromOutside = useCallback(
-    () => {
-      setMyEvents((prev) => {
-        const existing = draggedEvent;
-        console.log('existing', existing)
-        existing.in_calendar = !existing.in_calendar;
-        const filtered = prev.filter((ev) => ev.id !== draggedEvent.id);
-        return [...filtered, { ...existing }];
-      });
-      setDraggedEvent(null)
-    },
-    [draggedEvent, setDraggedEvent, newEvent]
-  )
 
   // const naviBar = (<TopBar isMobile={isMobile} onCalendar={onCalendar} setOnCalendar={setOnCalendar}/>)
   // const toDoList = (<ToDoList addToCalendar={addToCalendar}/>)
@@ -149,17 +50,6 @@ function App() {
       isMobile={isMobile}
       isLoggedIn={isLoggedIn}
       setIsLoggedIn={setIsLoggedIn}
-      onCalendar={onCalendar}
-      setOnCalendar={setOnCalendar}
-      myEvents={myEvents}
-      moveEvent={moveEvent}
-      resizeEvent={resizeEvent}
-      newEvent={newEvent}
-      changeTitle={changeTitle}
-      handleDragStart={handleDragStart}
-      draggedEvent={draggedEvent}
-      setDraggedEvent={setDraggedEvent}
-      onDropFromOutside={onDropFromOutside}
       isLoading={isLoading}
       userEmail={userEmail}
     />
