+++ client/src/components/Home.jsx
@@ -1,9 +1,10 @@
-import React, { useState, useEffect, useRef } from 'react';
+import React, { useState, useEffect, useCallback, useRef } from 'react';
 import { useNavigate, BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import axios from 'axios';
 import MyCalendar from './calendar/MyCalendar.jsx';
 import ToDoList from './to-do-list/ToDoList.jsx';
 import TopBar from './TopBar.jsx';
+import moment from 'moment';
 
 // For example data:
 import { result } from '../../../database/example.js';
@@ -14,10 +15,41 @@ import { result } from '../../../database/example.js';
   // const myCalender = (<MyCalendar myEvents={myEvents} moveEvent={moveEvent} resizeEvent={resizeEvent} changeTitle={changeTitle} onDropFromOutside={onDropFromOutside}/>)
   // const testToDo = (<TestToDo draggedEvent={draggedEvent} setDraggedEvent={setDraggedEvent} handleDragStart={handleDragStart}/>)
   // condition redering base on device
-const Home = ({isMobile, isLoggedIn, isLoading, setIsLoggedIn, userEmail, sharedBy, onCalendar, setOnCalendar,
-  myEvents, moveEvent, resizeEvent, changeTitle, handleDragStart, draggedEvent, setDraggedEvent, onDropFromOutside}) => {
+const Home = ({isMobile, isLoggedIn, isLoading, setIsLoggedIn, userEmail, sharedBy}) => {
 
   const [allTodos, setAllTodos] = useState([]);
+  const [myEvents, setMyEvents] = useState([
+    {
+      id: 0,
+      title: "Sample Event",
+      start_date: new Date(),
+      end_date: new Date(moment().add(1, "hour")),
+      allDay: false,
+      in_calendar: true
+    },
+    {
+      id: 1,
+      title: 'Trip to China',
+      description: '5-day business trip to meet with manufacturers',
+      duration: '5 days',
+      start: new Date(),
+      end: new Date(moment().add(5, "days")),
+      category_id: 1,
+      in_calendar: false
+    },
+    {
+      id: 2,
+      title: 'Trip to Los Angeles',
+      description: 'Meeting with Executives',
+      duration: '6 hours',
+      start: new Date(moment('14 Apr 2022 09:30:00 UT')),
+      end: new Date(moment('14 Apr 2022 15:30:00 UT')),
+      category_id: 1,
+      in_calendar: false
+    }
+  ]);
+  const [onCalendar, setOnCalendar] = useState(false);
+  const [draggedEvent, setDraggedEvent] = useState()
 
   // Requires logic and transforming of data to appropriate form to pass down data to calendar and todoList with required information from the appropriate user's calendar
   // All Components lose functionality if "isLoggedIn" is false
@@ -79,6 +111,72 @@ const Home = ({isMobile, isLoggedIn, isLoading, setIsLoggedIn, userEmail, shared
   // useEffect(() => {
   //   getAllTodos();
   // })
+  //Calendar helper functions
+  const moveEvent = useCallback(
+    ({ event, start, end, isAllDay: droppedOnAllDaySlot = false }) => {
+      const { allDay } = event;
+      if (!allDay && droppedOnAllDaySlot) {
+        event.allDay = true;
+      }
+
+      setMyEvents((prev) => {
+        const existing = prev.find((ev) => ev.id === event.id) ?? {};
+        const filtered = prev.filter((ev) => ev.id !== event.id);
+        return [...filtered, { ...existing, start, end, allDay }];
+      });
+    },
+    [setMyEvents]
+  );
+
+  const newEvent = useCallback(
+    (event) => {
+      setMyEvents((prev) => {
+        const idList = prev.map((item) => item.id)
+        const newId = Math.max(...idList) + 1
+        return [...prev, { ...event, id: newId }]
+      })
+    },
+    [setMyEvents]
+  )
+
+  const resizeEvent = useCallback(
+    ({ event, start, end }) => {
+      setMyEvents((prev) => {
+        const existing = prev.find((ev) => ev.id === event.id) ?? {};
+        const filtered = prev.filter((ev) => ev.id !== event.id);
+        return [...filtered, { ...existing, start, end }];
+      });
+    },
+    [setMyEvents]
+  );
+
+  const changeTitle = (event) => {
+    var title = prompt("Change title", event.title);
+    var newList = myEvents;
+    setMyEvents((prev) => {
+      newList[prev.indexOf(event)].title = title;
+      return newList
+    });
+  };
+  const handleDragStart = useCallback((event) => {
+    console.log('dragged event', event)
+    setDraggedEvent(event), []
+  })
+
+
+  const onDropFromOutside = useCallback(
+    () => {
+      setMyEvents((prev) => {
+        const existing = draggedEvent;
+        console.log('existing', existing)
+        existing.in_calendar = !existing.in_calendar;
+        const filtered = prev.filter((ev) => ev.id !== draggedEvent.id);
+        return [...filtered, { ...existing }];
+      });
+      setDraggedEvent(null)
+    },
+    [draggedEvent, setDraggedEvent, newEvent]
+  )
 
   // All Components
   const naviBar = (<TopBar isLoading={isLoading} setIsLoggedIn={setIsLoggedIn} isLoggedIn={isLoggedIn} isMobile={isMobile} onCalendar={onCalendar} setOnCalendar={setOnCalendar}/>);
